requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
httpx==0.27.2
python-dotenv==1.0.1
google-genai>=0.8.0
